Introduce new ai-node-sdk package
