+++ packages/@n8n/ai-node-sdk/.gitignore
@@ -1,0 +1,2 @@
+dist
+dist_examples
