+++ packages/@n8n/ai-node-sdk/eslint.config.mjs
@@ -1,16 +1,13 @@
 import { defineConfig, globalIgnores } from 'eslint/config';
 import { nodeConfig } from '@n8n/eslint-config/node';
 
-export default defineConfig(nodeConfig, globalIgnores(['scripts/**']), {
+export default defineConfig(nodeConfig, globalIgnores(['dist_examples/**']), {
 	rules: {
 		'@typescript-eslint/no-explicit-any': 'warn',
 		'@typescript-eslint/no-unsafe-assignment': 'warn',
 		'@typescript-eslint/no-unsafe-call': 'warn',
 		'@typescript-eslint/no-unsafe-member-access': 'warn',
 		'@typescript-eslint/no-unsafe-return': 'warn',
-		'no-case-declarations': 'warn',
-		'@typescript-eslint/require-await': 'warn',
-		'@typescript-eslint/prefer-nullish-coalescing': 'warn',
 		'@typescript-eslint/naming-convention': 'warn',
 	},
 });
