+++ packages/@n8n/ai-node-sdk/examples/nodes/LmChatOpenAiCustom/description.ts
