+++ packages/@n8n/ai-node-sdk/examples/nodes/LmChatOpenAiSimple/LmChatOpenAiSimple.node.ts
@@ -1,7 +1,6 @@
+import { createChatModelNode } from '@n8n/ai-utilities';
+import type { ProviderTool } from '@n8n/ai-utilities';
 import { NodeConnectionTypes, type IDataObject, type ISupplyDataFunctions } from 'n8n-workflow';
-
-import { createChatModelNode } from 'src/creators/create-chat-model-node';
-import type { ProviderTool } from 'src/types/tool';
 
 import { formatBuiltInTools } from '../common';
 import { openAiProperties } from '../properties';
@@ -10,7 +9,7 @@ export type ModelOptions = {
 	temperature?: number;
 };
 
-export class LmChatOpenAiSimple extends createChatModelNode({
+export const LmChatOpenAiSimple = createChatModelNode({
 	description: {
 		displayName: 'OpenAI Simple',
 
@@ -77,4 +76,4 @@ export class LmChatOpenAiSimple extends createChatModelNode({
 			providerTools,
 		};
 	},
-}) {}
+});
