+++ packages/@n8n/ai-node-sdk/examples/nodes/properties.ts
