+++ packages/@n8n/ai-node-sdk/integration-tests/openai.test.ts
@@ -1,3 +1,4 @@
+import { LangchainAdapter } from '@n8n/ai-utilities';
 import { createAgent, HumanMessage } from 'langchain';
 import nock from 'nock';
 
@@ -10,8 +11,7 @@ import {
 	mockToolCallResponse,
 	weatherTool,
 } from './openai.fixtures';
-import { OpenAIChatModel } from '../examples/models/openai';
-import { LangchainAdapter } from '../src/adapters/langchain-chat-model';
+import { OpenAIChatModel } from '../examples/models/openai/model';
 
 describe('OpenAI Integration with Langchain Agent', () => {
 	const baseURL = 'https://api.openai.com/v1';
