+++ packages/@n8n/ai-node-sdk/jest.config.js
@@ -1,0 +1,6 @@
+/** @type {import('jest').Config} */
+module.exports = {
+	...require('../../../jest.config'),
+	collectCoverageFrom: ['src/**/*.ts', 'integration-tests/**/*.ts'],
+	coveragePathIgnorePatterns: ['examples'],
+};
