+++ packages/@n8n/ai-node-sdk/src/index.ts
@@ -1,0 +1,46 @@
+export { logWrapper } from '@n8n/ai-utilities';
+export { parseSSEStream } from '@n8n/ai-utilities';
+export type { ChatModel, ChatModelConfig } from '@n8n/ai-utilities';
+export type { GenerateResult, StreamChunk, TokenUsage, FinishReason } from '@n8n/ai-utilities';
+export type { Tool, ToolResult, ToolCall, ProviderTool } from '@n8n/ai-utilities';
+export type {
+	Message,
+	ContentFile,
+	ContentMetadata,
+	ContentReasoning,
+	ContentText,
+	ContentToolCall,
+	ContentToolResult,
+	MessageContent,
+	MessageRole,
+} from '@n8n/ai-utilities';
+export type { JSONArray, JSONObject, JSONValue } from '@n8n/ai-utilities';
+export type { ServerSentEventMessage } from '@n8n/ai-utilities';
+export { BaseChatModel } from '@n8n/ai-utilities';
+export { getParametersJsonSchema } from '@n8n/ai-utilities';
+export type {
+	ChatModelNodeConfig,
+	ChatMemoryNodeConfig,
+	OpenAiModel,
+	ChatModelOptions,
+	GetChatModelFn,
+	GetChatMemoryFn,
+} from '@n8n/ai-utilities';
+
+// Memory types
+export type { ChatHistory, ChatMemory } from '@n8n/ai-utilities';
+
+// Memory base classes
+export { BaseChatHistory } from '@n8n/ai-utilities';
+export { BaseChatMemory } from '@n8n/ai-utilities';
+
+// Memory implementations
+export { WindowedChatMemory, type WindowedChatMemoryConfig } from '@n8n/ai-utilities';
+
+// Suppliers
+export { supplyMemory } from '@n8n/ai-utilities';
+export { supplyModel } from '@n8n/ai-utilities';
+
+// Creators
+export { createChatModelNode } from '@n8n/ai-utilities';
+export { createChatMemoryNode } from '@n8n/ai-utilities';
