+++ packages/@n8n/ai-node-sdk/tsconfig.examples.json
@@ -3,12 +3,9 @@
 	"compilerOptions": {
 		"composite": true,
 		"rootDir": ".",
-		"paths": {
-			"src/*": ["./src/*"],
-			"src": ["./src"]
-		},
 		"outDir": "./dist_examples",
 		"tsBuildInfoFile": "dist_examples/build.tsbuildinfo"
 	},
+	"include": ["src/**/*.ts", "examples/**/*.ts"],
 	"exclude": ["src/**/__tests__/**", "src/**/*.test.ts", "integration-tests/**/*.ts"]
 }
