+++ packages/@n8n/ai-utilities/jest.config.js
@@ -1,7 +1,6 @@
 /** @type {import('jest').Config} */
 module.exports = {
 	...require('../../../jest.config'),
-	collectCoverageFrom: ['src/**/*.ts', 'integration-tests/**/*.ts'],
+	collectCoverageFrom: ['src/**/*.ts'],
 	setupFilesAfterEnv: ['jest-expect-message'],
-	coveragePathIgnorePatterns: ['examples'],
 };
