+++ packages/@n8n/ai-utilities/package.json
@@ -6,12 +6,10 @@
 	"module": "src/index.ts",
 	"types": "dist/index.d.ts",
 	"scripts": {
-		"example:run": "tsx examples/run.ts",
 		"clean": "rimraf dist .turbo",
 		"dev": "pnpm run watch",
 		"typecheck": "tsc --noEmit",
 		"copy-tokenizer-json": "node scripts/copy-tokenizer-json.js .",
-		"build:examples": "tsc --build tsconfig.examples.json && tsc-alias -p tsconfig.examples.json && pnpm copy-tokenizer-json ./dist_examples/src/",
 		"build": "tsc --build tsconfig.build.json && tsc-alias -p tsconfig.build.json && pnpm copy-tokenizer-json",
 		"format": "biome format --write .",
 		"format:check": "biome ci .",
