+++ packages/@n8n/ai-utilities/src/creators/create-chat-memory-node.ts
@@ -1,16 +1,16 @@
 import type { INodeType, ISupplyDataFunctions } from 'n8n-workflow';
 
 import { supplyMemory } from 'src/suppliers/supplyMemory';
-import type { ChatMemoryNodeConfig, UnextendableNodeType } from 'src/types/creators';
+import type { ChatMemoryNodeConfig } from 'src/types/creators';
 
 export const createChatMemoryNode = (chatMemoryNode: ChatMemoryNodeConfig) => {
-	const constructor = class ChatMemoryNode implements INodeType {
-		description = chatMemoryNode.description;
-		methods = chatMemoryNode.methods;
+	const node: INodeType = {
+		description: chatMemoryNode.description,
+		methods: chatMemoryNode.methods,
 		async supplyData(this: ISupplyDataFunctions, itemIndex: number) {
 			const memory = await chatMemoryNode.getMemory(this, itemIndex);
 			return supplyMemory(this, memory, chatMemoryNode.memoryOptions);
-		}
+		},
 	};
-	return constructor as UnextendableNodeType;
+	return node;
 };
