+++ packages/@n8n/ai-utilities/tsconfig.json
@@ -13,6 +13,6 @@
 		"emitDecoratorMetadata": true,
 		"experimentalDecorators": true
 	},
-	"include": ["src/**/*.ts", "test/**/*.ts", "examples/**/*.ts", "integration-tests/**/*.ts"],
+	"include": ["src/**/*.ts", "test/**/*.ts"],
 	"references": [{ "path": "../../workflow/tsconfig.build.esm.json" }]
 }
