+++ packages/@n8n/benchmark/package.json
@@ -35,7 +35,7 @@
   "dependencies": {
     "@oclif/core": "4.0.7",
     "axios": "catalog:",
-    "dotenv": "17.2.3",
+    "dotenv": "catalog:",
     "nanoid": "catalog:",
     "zx": "^8.1.4"
   },
