+++ packages/cli/package.json
@@ -134,7 +134,7 @@
     "convict": "6.2.4",
     "cookie-parser": "1.4.7",
     "csrf": "3.1.0",
-    "dotenv": "17.2.3",
+    "dotenv": "catalog:",
     "express": "5.1.0",
     "express-handlebars": "8.0.1",
     "express-openapi-validator": "5.5.3",
